gcc14: support Xcode 16
Xcode 16 `as` is a wrapper around `clang -x assembler`. Where previously
(through Xcode 15.x), `as` would read its input from standard input when
not invoked with an input path, Xcode 16 `as` will instead print “clang:
error: no input files” to standard error and exit with status 1.

This makes it impossible to use `gcc -pipe`. MacPorts specifies `-pipe`
by default. (…/etc/macports/macports.conf configurepipe)

Xcode 16 `as` will read its input from standard input if invoked with
`-` as the input path. Previous versions of `as`, including Xcode 15
`as`, respect this convention as well, as should binutils `gas`.

gcc has a macro, AS_NEEDS_DASH_FOR_PIPED_INPUT, which can be defined to
configure it to operate with `as` programs that require this convention.
This change enables that option for darwin-gcc.